BUNDLE: avoid crash on cleanup when checkpoint is deactivated.

+++ BUNDLE/BENCH_C016_02/DATA/cs_user_scripts.py
@@ -38,8 +38,9 @@ def domain_copy_results_add(domain):
 
     # Clean some temporary files
 
-    import shutil
-    shutil.rmtree('checkpoint')
+    if os.path.exists('checkpoint'):
+        import shutil
+        shutil.rmtree('checkpoint')
 
     return
